+++ src/components/UnitLinkList.tsx
@@ -1,16 +1,25 @@
-import {LinkList} from "../components/LinkList";
+import {LinkList, LinkListItem} from "../components/LinkList";
 import {linkHTML, LinkTarget} from "../db/link-target";
 import {toTagString} from "../tag";
 
 export interface UnitLinkListProps {
     title?: string,
+    depth?: number,
     items: LinkTarget[],
 }
 
+
+function toLinkListItem(target: LinkTarget, depth: number): LinkListItem {
+    return {
+        content: linkHTML(target),
+        href: `/t/${toTagString(target.tag)}`,
+        children: depth > 0 && target.children && target.children.length > 0 ?
+            target.children.map((c) => toLinkListItem(c, depth - 1)) : []
+    };
+}
+
+
 export function UnitLinkList(props: UnitLinkListProps) {
-    return <LinkList title={props.title} items={props.items.map((t) => ({
-        content: linkHTML(t),
-        href: `/t/${toTagString(t.tag)}`
-    }))}/>
+    return <LinkList title={props.title} items={props.items.map((t) => toLinkListItem(t, props.depth ?? 0))}/>
 }