+++ changelog.md
@@ -6,6 +6,7 @@
 - Added font size, line height, and alignment config options.
 - Units now get an automatically generated label based on its numbering if none is given. 
 - Added the conservative mode to the `watch` command.
+- Added config option to adjust the depth of the table of contents. 
 
 
 # v0.1.6, 2026-03-19